Add linting to CI
The workflow will trigger on PRs targeting `main` and on pushes
that are not pushes on `main`.

Using the `slim` version to gain some speed compared to the full
version of super-linter.

Validation only runs on files that are changed and not the entire
codebase.

We will probably need to configure the different linters, but that
can be part of a separate commit.

See: #19