Start with backend development
Creating a dev docker image based on python3.9 and poetry.
Fixing poetry version for reproducible docker builds.
Because we are inside a container poetry doesn't need to use
virtual environments. Pre-installing our python dependencies, so
containers based on the image are ready to go for dev work.

Creating a docker-compose file to spin up a container from our dev
image. Mounting in `backend` to `/usr/src/aof` so changes are synced.
Using an infinite loop for the time being to keep the container
running. (Will change to start and serve the backend.)

Running poetry init with installing fastapi and uvicorn[standard] as
dependencies.

See: #17