Bump super-linter from 4.8.1 to 4.*
I believe there were some fixes introduced in 4.8.4 for mypy.

Ideally I would prefer to fix the minor version, but super-linter
does not publish tags for it. Only patch and major tags.

Instead of fixing and manually upping the version I think it's ok
to pin to the major version in this case.

See: #19