Update super-linter configuration
Configure isort to play nice with black. It also needs help in telling
the difference between 1st and 3rd party packages within super-linter.

Configure mypy to ignore missing imports as super-linter is not aware
of the project's dependencies. Checking for untyped decorators also
have to be disabled because of the decorators we use from fastapi.

We don't actually need to run the linter on pushes. (PR is enough.)

Close: #19