Update keycloak config
Rename `alphafold-on-fire` to `alphafold-frontend`
Add new client `alphafold-backend`, type `bearer-only`
Add role `authenticated` to alphafold-backend
Make the role an automatic default role for every user in keycloak
(Under roles -> default roles -> client roles -> alphafold-backend)

When a new auth request comes in for any client - if the client returns
a token with full scopes - the access token will contain the new role
for the user and `aud` will be populated with `alphafold-backend`.
The value of `aud` becomes a list.
e.g. `["alphafold-backend","account"]`