Secure backend
security.py
`security.get_claims()` can be used as a dependency in the routes.
Fastapi reject request if there is no bearer token (`_oauth_scheme`).
Get `iss` from token -> get jwks from auth server -> validate token,
if anything goes wrong reject request.
We just need to use `get_claims()` on the routes we want to protect.

config.py
Use `pydantic.AnyHttpUrl` to allow using `localhost`.
(No top-level-domain validation.)

test-backend.yaml
Needed a re-work to be able to use a keycloak instance for testing.
The realm configuration is loaded from the workspace, hence we cannot
use the services section for keycloak. It is also slow to start-up, so
had to build in an extra step to validate it's up-n-running.

docker-compose.yaml
Because both the frontend (swagger-ui) and the backend (fastapi) has to
talk to keycloak. I had to get the IP address of the keycloak container
when it's running and map that to `localhost` inside the app container.
This way the localhost auth server url works in both places.